Extract useCurrentRelease hook from DeployTab
Moves the listCurrentReleases query and the per-environment filter
memo into a dedicated hook, completing the hook extraction series.
